# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeGuard AI API — emergency session management, contacts, AI chat, GPS broadcasts
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EmergencySessionTriggerType = typeof EmergencySessionTriggerType[keyof typeof EmergencySessionTriggerType];


export const EmergencySessionTriggerType = {
  fall: 'fall',
  crash: 'crash',
  manual: 'manual',
  sos: 'sos',
} as const;

export type EmergencySessionStatus = typeof EmergencySessionStatus[keyof typeof EmergencySessionStatus];


export const EmergencySessionStatus = {
  pending: 'pending',
  active: 'active',
  cancelled: 'cancelled',
  resolved: 'resolved',
} as const;

export interface EmergencySession {
  id: number;
  triggerType: EmergencySessionTriggerType;
  status: EmergencySessionStatus;
  latitude: number;
  longitude: number;
  /** @nullable */
  accelerometerPeak?: number | null;
  /** @nullable */
  countdownSeconds?: number | null;
  /** @nullable */
  resolvedAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type EmergencySessionInputTriggerType = typeof EmergencySessionInputTriggerType[keyof typeof EmergencySessionInputTriggerType];


export const EmergencySessionInputTriggerType = {
  fall: 'fall',
  crash: 'crash',
  manual: 'manual',
  sos: 'sos',
} as const;

export interface EmergencySessionInput {
  triggerType: EmergencySessionInputTriggerType;
  latitude: number;
  longitude: number;
  /** @nullable */
  accelerometerPeak?: number | null;
  countdownSeconds?: number;
  /** @nullable */
  notes?: string | null;
}

export type EmergencySessionUpdateStatus = typeof EmergencySessionUpdateStatus[keyof typeof EmergencySessionUpdateStatus];


export const EmergencySessionUpdateStatus = {
  pending: 'pending',
  active: 'active',
  cancelled: 'cancelled',
  resolved: 'resolved',
} as const;

export interface EmergencySessionUpdate {
  status?: EmergencySessionUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export interface Contact {
  id: number;
  name: string;
  phone: string;
  relationship: string;
  isPrimary?: boolean;
  createdAt: string;
}

export interface ContactInput {
  name: string;
  phone: string;
  relationship: string;
  isPrimary?: boolean;
}

export interface ContactUpdate {
  name?: string;
  phone?: string;
  relationship?: string;
  isPrimary?: boolean;
}

export interface ChatSession {
  id: number;
  title: string;
  /**
     * Optional context tag (emergency, general, firstaid)
     * @nullable
     */
  context?: string | null;
  messageCount?: number;
  createdAt: string;
}

export interface ChatSessionInput {
  title: string;
  /** @nullable */
  context?: string | null;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: number;
  sessionId: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface ChatMessageInput {
  content: string;
}

export interface GpsBroadcast {
  id: number;
  /** @nullable */
  sessionId?: number | null;
  latitude: number;
  longitude: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  speed?: number | null;
  /** @nullable */
  altitude?: number | null;
  createdAt: string;
}

export interface GpsBroadcastInput {
  /** @nullable */
  sessionId?: number | null;
  latitude: number;
  longitude: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  speed?: number | null;
  /** @nullable */
  altitude?: number | null;
}

export type StatsLastBroadcastLocation = {
  latitude?: number;
  longitude?: number;
  timestamp?: string;
} | null;

export interface Stats {
  totalEmergencies: number;
  totalContacts: number;
  resolvedEmergencies: number;
  cancelledEmergencies: number;
  recentEmergencies: EmergencySession[];
  lastBroadcastLocation?: StatsLastBroadcastLocation;
}

export type ListEmergencySessionsParams = {
limit?: number;
status?: ListEmergencySessionsStatus;
};

export type ListEmergencySessionsStatus = typeof ListEmergencySessionsStatus[keyof typeof ListEmergencySessionsStatus];


export const ListEmergencySessionsStatus = {
  pending: 'pending',
  active: 'active',
  cancelled: 'cancelled',
  resolved: 'resolved',
} as const;

export type ListGpsBroadcastsParams = {
/**
 * @nullable
 */
sessionId?: number | null;
limit?: number;
};