# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeGuard AI API — emergency session management, contacts, AI chat, GPS broadcasts
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageInput {
  content: string;
}