# lib/api-zod/src/generated/types/gpsBroadcast.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeGuard AI API — emergency session management, contacts, AI chat, GPS broadcasts
 * OpenAPI spec version: 0.1.0
 */

export interface GpsBroadcast {
  id: number;
  /** @nullable */
  sessionId?: number | null;
  latitude: number;
  longitude: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  speed?: number | null;
  /** @nullable */
  altitude?: number | null;
  createdAt: string;
}