# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeGuard AI API — emergency session management, contacts, AI chat, GPS broadcasts
 * OpenAPI spec version: 0.1.0
 */
import type { EmergencySession } from './emergencySession';
import type { StatsLastBroadcastLocation } from './statsLastBroadcastLocation';

export interface Stats {
  totalEmergencies: number;
  totalContacts: number;
  resolvedEmergencies: number;
  cancelledEmergencies: number;
  recentEmergencies: EmergencySession[];
  lastBroadcastLocation?: StatsLastBroadcastLocation;
}